This branch includes classes which have been moved to individual files. This does not compile.
